DP-5994 - Retry query request after database restart
Connections are stored in a LRU cache. But if the database is restarted in the meantime
the connection to that database is invalid. iThe query will fail and the connection is closed.
If the the query is retried a new connection will be created and the request will be succesfull

This is implemented as a decorator that catches the exception and handles the retries