AB#26211 Use path to construct urls for dso datasets
For DSO datasets, the geosearch is looking into `datasets_dataset` to
grab the name for a dataset, that will be used in the construction
of resulting urls.

However, for DSO datasets with extra path components, this will results
in wrong urls. Luckily, the `datasets_dataset.path` column is
already available and can be used to generate the correct urls.

Because non-DSO datasets do not have a `path` column,
for these datasets the original `name` column is still being used.