Exclude non-active records for datasets with historical data
For the geosearch, when searching in datasets with historical
records (temporality), we are only interested in the currently active records.

The geosearch now collects the schema (for amsterdam schema based
datasets) from the database, and determines if the dataset
has temporality. The temporality information is used
to add extra search clauses to the SQL queries.