AB#78345 Add possibility to have extra field in the response.
NB. This only works for amsterdam schema based datasets,
because in that case we have field information available.

Normally, geosearch returns a fixed set a fields (with some variation
per search backend). However, in some situations it can make querying
much more efficient if extra fields can be returned immediately,
because no round trips to additional APIs is needed.

The DSO-API standard has been followed, where a `_fields` param
can be added to the requesting querystring that holds a comma-separated set of fields
(snake_cased or camelCased).

If those fieldnames are matching with the available fields on the
datasettable, the response will include those fields and their values.

Because we are only handling amsterdam schema based datasets,
the fields will we be camelcased in the output.