Fix extra `_fields` feature to work with more datasets
The `_fields` option was not working for all datasets.
For names with underscores, the dataset query class was not initialized
propertly. Because the tablename in `datasets_datasettable` is
snakecased, but the actual table id is camelcased.

 And for schemas that are split up over multiple files,
the query class was also not initialized. So, the schematools class loader
needed to be introduced to fix this problem.