Fix versioned tables
Geosearch is using a query on the datasets tables to fill it's registry.
This could possibly fail on the versioned tables, since there was no
logic for this scenario. A quick fix has been implemented to always get
the default version.

Tests were also failing because of the order in which records were
returned from the database. As a quick fix the order of inserts is
changed, but we need to look into a permanent solution or refactor to
solve the underlying issues.